Disable multicast scouting of CLI in testing
This is to ensure we're connecting to right router during tests,
as Cargo runs tests in-parallel, there could be multiple Zenoh routers running concurrently.

Just providing `connect.endpoints` is not sufficient to connect to a single router,
due to multicast scouting the testing CLIs could still connect to other testing routers.

Disabling this ensures the testing CLIs always connects to the corresponding testing routers.